Change mousewheel behavior for the timeline flamechart
