+++ src/content/components/FlameChartCanvas.js
@@ -3,6 +3,7 @@ import React, { Component } from 'react';
 import shallowCompare from 'react-addons-shallow-compare';
 import { timeCode } from '../../common/time-code';
 import TextMeasurement from '../../common/text-measurement';
+import withTimelineViewport from './TimelineViewport';
 
 import type { Thread } from '../../common/types/profile';
 import type { Milliseconds, CssPixels, UnitIntervalOfProfileRange, DevicePixels } from '../../common/types/units';
@@ -102,7 +103,7 @@ class FlameChartCanvas extends Component {
    *
    * Note that most of the units are not absolute values, but unit intervals ranged from
    * 0 - 1. This was done to make the calculations easier for computing various zoomed
-   * and translated views independent of any particular scale. See FlameChartViewport.js
+   * and translated views independent of any particular scale. See TimelineViewport.js
    * for a diagram detailing the various components of this set-up.
    * @param {HTMLCanvasElement} canvas - The current canvas.
    * @returns {undefined}
@@ -191,4 +192,4 @@ class FlameChartCanvas extends Component {
   }
 }
 
-export default FlameChartCanvas;
+export default withTimelineViewport(FlameChartCanvas);
