--- src/content/components/FlameChartViewport.css
@@ -1,37 +1,0 @@
-.flameChartViewport {
-  flex: 1;
-  overflow: auto;
-  position: relative;
-  margin-top: 0;
-  transition: transform 100ms;
-}
-
-.flameChartViewport.expanded {
-  cursor: -webkit-grab;
-  cursor: grab;
-  border-top: 1px solid #d6d6d6;
-  border-bottom: 1px solid #d6d6d6;
-  margin-bottom: 5px;
-}
-
-.flameChartViewport.expanded::after {
-  content: "";
-  pointer-events: none;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  /* Slight blue shadow. */
-  box-shadow: inset 0 0 20px rgba(64, 115, 140, 0.2);
-  z-index:1;
-}
-
-.flameChartViewport.collapsed {
-  transform: translateY(9px);
-}
-
-.flameChartViewport.dragging {
-  cursor: -webkit-grabbing;
-  cursor: grabbing;
-}
